move queue to separate file
