+++ panther_lights/src/controller_node.cpp
@@ -31,6 +31,7 @@
 
 #include <panther_msgs/srv/set_led_animation.hpp>
 
+#include <panther_lights/led_animations_queue.hpp>
 #include <panther_lights/led_panel.hpp>
 #include <panther_lights/led_segment.hpp>
 #include <panther_lights/segment_converter.hpp>
@@ -41,122 +42,6 @@ namespace panther_lights
 
 using std::placeholders::_1;
 using std::placeholders::_2;
-
-LEDAnimation::LEDAnimation(
-  const LEDAnimationDescription & led_animation_description,
-  const std::unordered_map<std::string, std::shared_ptr<LEDSegment>> & segments,
-  const rclcpp::Time & init_time)
-: led_animation_description_(led_animation_description),
-  init_time_(init_time),
-  repeating_(false),
-  param_("")
-{
-  for (auto & animation : led_animation_description_.animations) {
-    for (auto & segment : animation.segments) {
-      animation_segments_.push_back(segments.at(segment));
-    }
-  }
-}
-
-bool LEDAnimation::IsFinished()
-{
-  return std::all_of(
-    animation_segments_.begin(), animation_segments_.end(),
-    [](const std::shared_ptr<LEDSegment> & segment) { return segment->IsAnimationFinished(); });
-}
-
-float LEDAnimation::GetProgress() const
-{
-  float progress = 1.0f;
-  std::for_each(
-    animation_segments_.begin(), animation_segments_.end(),
-    [&progress](const std::shared_ptr<LEDSegment> & segment) {
-      auto anim_progress = segment->GetAnimationProgress();
-      progress = anim_progress < progress ? anim_progress : progress;
-    });
-
-  return progress;
-}
-
-void LEDAnimation::Reset() const
-{
-  std::for_each(
-    animation_segments_.begin(), animation_segments_.end(),
-    [](const std::shared_ptr<LEDSegment> & segment) { segment->ResetAnimation(); });
-}
-
-void AnimationsQueue::Put(
-  const std::shared_ptr<LEDAnimation> & animation, const rclcpp::Time & time)
-{
-  if (animation->GetPriority() == 1) {
-    Clear();
-  }
-  Validate(time);
-
-  if (queue_.size() == max_queue_size_) {
-    throw std::runtime_error("Animation queue overloaded");
-  }
-
-  queue_.push_back(animation);
-  std::sort(
-    queue_.begin(), queue_.end(),
-    [](const std::shared_ptr<LEDAnimation> & a, const std::shared_ptr<LEDAnimation> & b) {
-      return a->GetPriority() < b->GetPriority() ||
-             (a->GetPriority() == b->GetPriority() && a->GetInitTime() < b->GetInitTime());
-    });
-}
-
-std::shared_ptr<LEDAnimation> AnimationsQueue::Get()
-{
-  if (!queue_.empty()) {
-    const auto animation = queue_.front();
-    queue_.pop_front();
-    return animation;
-  }
-  throw std::runtime_error("Queue empty");
-}
-
-void AnimationsQueue::Clear(std::size_t priority)
-{
-  const auto new_end = std::remove_if(
-    queue_.begin(), queue_.end(), [priority](const std::shared_ptr<LEDAnimation> & animation) {
-      return animation->GetPriority() >= priority;
-    });
-  queue_.erase(new_end, queue_.end());
-}
-
-void AnimationsQueue::Remove(const std::shared_ptr<LEDAnimation> & animation)
-{
-  auto it = std::find(queue_.begin(), queue_.end(), animation);
-  if (it != queue_.end()) {
-    queue_.erase(it);
-  }
-}
-
-bool AnimationsQueue::HasAnimation(const std::shared_ptr<LEDAnimation> & animation) const
-{
-  return std::find(queue_.begin(), queue_.end(), animation) != queue_.end();
-}
-
-void AnimationsQueue::Validate(const rclcpp::Time & time)
-{
-  for (auto it = queue_.begin(); it != queue_.end();) {
-    if ((time - it->get()->GetInitTime()).seconds() > it->get()->GetTimeout()) {
-      // TODO: log info - animation timeout
-      it = queue_.erase(it);
-    } else {
-      ++it;
-    }
-  }
-}
-
-std::size_t AnimationsQueue::GetFirstAnimationPriority() const
-{
-  if (!Empty()) {
-    return queue_.front()->GetPriority();
-  }
-  return LEDAnimation::kDefaultPriority;
-}
 
 ControllerNode::ControllerNode(const std::string & node_name, const rclcpp::NodeOptions & options)
 : Node(node_name, options)
@@ -178,7 +63,7 @@ ControllerNode::ControllerNode(const std::string & node_name, const rclcpp::Node
 
   segment_converter_ = std::make_shared<SegmentConverter>();
 
-  animation_queue_ = std::make_unique<AnimationsQueue>(10);
+  animations_queue_ = std::make_unique<LEDAnimationsQueue>(10);
 
   set_led_animation_server_ = this->create_service<SetLEDAnimationSrv>(
     "lights/controller/set/animation", std::bind(&ControllerNode::SetLEDAnimationCB, this, _1, _2));
@@ -347,11 +232,11 @@ void ControllerNode::ControllerTimerCB()
   auto brightness = 255;
 
   if (animation_finished_) {
-    animation_queue_->Validate(this->get_clock()->now());
+    animations_queue_->Validate(this->get_clock()->now());
 
-    if (!animation_queue_->Empty()) {
+    if (!animations_queue_->Empty()) {
       try {
-        SetLEDAnimation(animation_queue_->Get());
+        SetLEDAnimation(animations_queue_->Get());
       } catch (const std::runtime_error & e) {
         RCLCPP_ERROR(this->get_logger(), "Failed to Set LED animation: %s", e.what());
       }
@@ -364,10 +249,10 @@ void ControllerNode::ControllerTimerCB()
     return;
   }
 
-  if (current_animation_->GetPriority() > animation_queue_->GetFirstAnimationPriority()) {
+  if (current_animation_->GetPriority() > animations_queue_->GetFirstAnimationPriority()) {
     if (current_animation_->GetProgress() < 0.65f) {
       current_animation_->Reset();
-      animation_queue_->Put(current_animation_, this->get_clock()->now());
+      animations_queue_->Put(current_animation_, this->get_clock()->now());
     }
     animation_finished_ = true;
     return;
@@ -415,8 +300,8 @@ void ControllerNode::AddAnimationToQueue(const std::size_t animation_id, const b
     animation_description, segments_, this->get_clock()->now());
   animation->SetRepeating(repeating);
   // animation->SetParam(param);
-  animation_queue_->Put(animation, this->get_clock()->now());
-  animation_queue_->Print();
+  animations_queue_->Put(animation, this->get_clock()->now());
+  animations_queue_->Print();
 }
 
 void ControllerNode::SetLEDAnimation(const std::shared_ptr<LEDAnimation> & led_animation)
